+++ tests/GoodReadsTest.php
@@ -23,6 +23,180 @@ class GoodReadsTest extends TestCase
         $_SERVER['SCRIPT_NAME'] = '/phpunit';
     }
 
+    /**
+     * Summary of testAppGetAuthors
+     * @return void
+     */
+    #[\PHPUnit\Framework\Attributes\RunInSeparateProcess]
+    public function testAppGetAuthors(): void
+    {
+        $_SERVER['PATH_INFO'] = '/gr_author/0';
+
+        ob_start();
+        $headers = headers_list();
+        require dirname(__DIR__) . '/app/index.php';
+        $output = ob_get_clean();
+
+        $expected = '<title>EPub Loader</title>';
+        $this->assertStringContainsString($expected, $output);
+        $expected = '<a href="/phpunit/gr_author/0">Some Books</a>';
+        $this->assertStringContainsString($expected, $output);
+        $expected = 'Arthur Conan Doyle';
+        $this->assertStringContainsString($expected, $output);
+
+        unset($_SERVER['PATH_INFO']);
+    }
+
+    /**
+     * Summary of testAppGetAuthorLinks
+     * @return void
+     */
+    #[\PHPUnit\Framework\Attributes\RunInSeparateProcess]
+    public function testAppGetAuthorLinks(): void
+    {
+        $_SERVER['PATH_INFO'] = '/gr_author/0';
+        $_GET['findLinks'] = '1';
+
+        ob_start();
+        $headers = headers_list();
+        require dirname(__DIR__) . '/app/index.php';
+        $output = ob_get_clean();
+
+        $expected = '<title>EPub Loader</title>';
+        $this->assertStringContainsString($expected, $output);
+        $expected = '<a href="/phpunit/gr_author/0/4?matchId=880695.H_G_Wells">880695.H_G_Wells</a>';
+        $this->assertStringContainsString($expected, $output);
+        $expected = 'H. G. Wells';
+        $this->assertStringContainsString($expected, $output);
+
+        unset($_SERVER['PATH_INFO']);
+        unset($_GET['findLinks']);
+    }
+
+    /**
+     * Summary of testAppGetAuthor
+     * @return void
+     */
+    #[\PHPUnit\Framework\Attributes\RunInSeparateProcess]
+    public function testAppGetAuthor(): void
+    {
+        $_SERVER['PATH_INFO'] = '/gr_author/0/1';
+
+        ob_start();
+        $headers = headers_list();
+        require dirname(__DIR__) . '/app/index.php';
+        $output = ob_get_clean();
+
+        $expected = '<title>EPub Loader</title>';
+        $this->assertStringContainsString($expected, $output);
+        $expected = '<a href="/phpunit/gr_books/0/1?authId=2448.Arthur_Conan_Doyle">2448.Arthur_Conan_Doyle</a>';
+        $this->assertStringContainsString($expected, $output);
+        $expected = 'Arthur Conan Doyle';
+        $this->assertStringContainsString($expected, $output);
+
+        unset($_SERVER['PATH_INFO']);
+    }
+
+    /**
+     * Summary of testAppGetBooks
+     * @return void
+     */
+    #[\PHPUnit\Framework\Attributes\RunInSeparateProcess]
+    public function testAppGetBooks(): void
+    {
+        $_SERVER['PATH_INFO'] = '/gr_books/0/1';
+
+        ob_start();
+        $headers = headers_list();
+        require dirname(__DIR__) . '/app/index.php';
+        $output = ob_get_clean();
+
+        $expected = '<title>EPub Loader</title>';
+        $this->assertStringContainsString($expected, $output);
+        $expected = '<a href="/phpunit/gr_books/0/1?bookId=11">Search</a>';
+        $this->assertStringContainsString($expected, $output);
+        $expected = 'A Study in Scarlet';
+        $this->assertStringContainsString($expected, $output);
+
+        unset($_SERVER['PATH_INFO']);
+    }
+
+    /**
+     * Summary of testAppGetBookSearch
+     * @return void
+     */
+    #[\PHPUnit\Framework\Attributes\RunInSeparateProcess]
+    public function testAppGetBookSearch(): void
+    {
+        $_SERVER['PATH_INFO'] = '/gr_books/0/1';
+        $_GET['bookId'] = '11';
+
+        ob_start();
+        $headers = headers_list();
+        require dirname(__DIR__) . '/app/index.php';
+        $output = ob_get_clean();
+
+        $expected = '<title>EPub Loader</title>';
+        $this->assertStringContainsString($expected, $output);
+        $expected = '<a href="/phpunit/gr_books/0/1?bookId=11&matchId=102868">102868</a>';
+        $this->assertStringContainsString($expected, $output);
+        $expected = 'A Study in Scarlet (Sherlock Holmes, #1)';
+        $this->assertStringContainsString($expected, $output);
+
+        unset($_SERVER['PATH_INFO']);
+        unset($_GET['bookId']);
+    }
+
+    /**
+     * Summary of testAppGetSeries
+     * @return void
+     */
+    #[\PHPUnit\Framework\Attributes\RunInSeparateProcess]
+    public function testAppGetSeries(): void
+    {
+        $_SERVER['PATH_INFO'] = '/gr_series/0';
+
+        ob_start();
+        $headers = headers_list();
+        require dirname(__DIR__) . '/app/index.php';
+        $output = ob_get_clean();
+
+        $expected = '<title>EPub Loader</title>';
+        $this->assertStringContainsString($expected, $output);
+        $expected = '<a href="/phpunit/gr_series/0/1?seriesId=1">Search</a>';
+        $this->assertStringContainsString($expected, $output);
+        $expected = 'Sherlock Holmes';
+        $this->assertStringContainsString($expected, $output);
+
+        unset($_SERVER['PATH_INFO']);
+    }
+
+    /**
+     * Summary of testAppGetSeriesMatch
+     * @return void
+     */
+    #[\PHPUnit\Framework\Attributes\RunInSeparateProcess]
+    public function testAppGetSeriesMatch(): void
+    {
+        $_SERVER['PATH_INFO'] = '/gr_series/0';
+        $_GET['matchId'] = '49996-sherlock-holmes';
+
+        ob_start();
+        $headers = headers_list();
+        require dirname(__DIR__) . '/app/index.php';
+        $output = ob_get_clean();
+
+        $expected = '<title>EPub Loader</title>';
+        $this->assertStringContainsString($expected, $output);
+        $expected = '<a href="/phpunit/gr_series/0/?matchId=49996-sherlock-holmes">49996-sherlock-holmes</a>';
+        $this->assertStringContainsString($expected, $output);
+        $expected = 'Sherlock Holmes is a fictional consulting detective in London';
+        $this->assertStringContainsString($expected, $output);
+
+        unset($_SERVER['PATH_INFO']);
+        unset($_GET['seriesId']);
+    }
+
     public function testJsonImportFile(): void
     {
         $dbPath = dirname(__DIR__) . '/cache/goodreads';
@@ -216,7 +390,7 @@ class GoodReadsTest extends TestCase
     public function testMatchParseAuthorResult(): void
     {
         $cacheDir = dirname(__DIR__) . '/cache';
-        //$match = new GoodReadsMatch($cacheDir);
+        $match = new GoodReadsMatch($cacheDir);
 
         $fileList = JsonImport::getFiles($cacheDir . '/goodreads/author/list/', '*.json');
         foreach ($fileList as $cacheFile) {
@@ -228,14 +402,14 @@ class GoodReadsTest extends TestCase
             $authors = GoodReadsBook::parseSearch($matched);
         }
 
-        $expected = 18;
+        $expected = count($match->getAuthorIds());
         $this->assertCount($expected, $fileList);
     }
 
     public function testMatchParseSearchResult(): void
     {
         $cacheDir = dirname(__DIR__) . '/cache';
-        //$match = new GoodReadsMatch($cacheDir);
+        $match = new GoodReadsMatch($cacheDir);
 
         $fileList = JsonImport::getFiles($cacheDir . '/goodreads/search/', '*.json');
         foreach ($fileList as $cacheFile) {
@@ -247,14 +421,14 @@ class GoodReadsTest extends TestCase
             $authors = GoodReadsBook::parseSearch($matched);
         }
 
-        $expected = 15;
+        $expected = count($match->getSearchQueries());
         $this->assertCount($expected, $fileList);
     }
 
     public function testMatchParseSeriesResult(): void
     {
         $cacheDir = dirname(__DIR__) . '/cache';
-        //$match = new GoodReadsMatch($cacheDir);
+        $match = new GoodReadsMatch($cacheDir);
 
         $fileList = JsonImport::getFiles($cacheDir . '/goodreads/series/', '*.json');
         foreach ($fileList as $cacheFile) {
@@ -267,14 +441,14 @@ class GoodReadsTest extends TestCase
             $series->setId($seriesId);
         }
 
-        $expected = 5;
+        $expected = count($match->getSeriesIds());
         $this->assertCount($expected, $fileList);
     }
 
     public function testMatchParseBook(): void
     {
         $cacheDir = dirname(__DIR__) . '/cache';
-        //$match = new GoodReadsMatch($cacheDir);
+        $match = new GoodReadsMatch($cacheDir);
 
         $fileList = JsonImport::getFiles($cacheDir . '/goodreads/book/show/', '*.json');
         foreach ($fileList as $cacheFile) {
@@ -286,7 +460,7 @@ class GoodReadsTest extends TestCase
             $book = GoodReadsBook::parse($matched);
         }
 
-        $expected = 4;
+        $expected = count($match->getBookIds());
         $this->assertCount($expected, $fileList);
     }
 }